requirements.txt
django-categories